requirements.txt
annotated-types==0.7.0
anyio==4.4.0
asgiref==3.8.1
Brotli==1.1.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.3.2
cssselect2==0.7.0
distro==1.9.0
Django==5.1.1
django-cors-headers==4.4.0
django-environ==0.11.2
djangorestframework==3.15.2
djangorestframework_simplejwt==5.5.0
et_xmlfile==2.0.0
fonttools==4.56.0
gunicorn==23.0.0
h11==0.14.0
httpcore==1.0.5
httpx==0.27.2
idna==3.10
jiter==0.5.0
numpy==2.2.3
openai==1.45.0
openpyxl==3.1.5
packaging==24.1
pandas==2.2.3
pillow==11.1.0
psycopg2-binary==2.9.9
pycparser==2.22
pydantic==2.9.1
pydantic_core==2.23.3
pydyf==0.11.0
PyJWT==2.9.0
pyphen==0.17.2
python-dateutil==2.9.0.post0
pytz==2025.1
requests==2.32.3
six==1.17.0
sniffio==1.3.1
sqlparse==0.5.1
tinycss2==1.4.0
tinyhtml5==2.0.0
tqdm==4.66.5
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.2.3
weasyprint==64.0
webencodings==0.5.1
zopfli==0.2.3.post1
